migrate proposal test
